+++ source/ftrack_location_compatibility/plugin.py
@@ -195,7 +195,14 @@ class ProxyLegacyLocationMixin(object):
 
 def register_locations(session):
     '''Register proxy locations.'''
-    for location in session.query('select name from Location').all():
+    for location in session.query('select name from Location')
+
+        '''
+        Since the implementation in the new API does not
+        translate paths between operating systems using ftrack disks
+        we mixin the proxy location with all the locatino which requires such a feature.
+        '''
+
         if (
             location.accessor and
             location['id'] is not ftrack_api.symbol.UNMANAGED_LOCATION_ID