+++ doc/release/release_notes.rst
@@ -11,8 +11,9 @@ Release Notes
     
     .. change:: new
 
-        Proxy the unmanaged location using the legacy API, since the implementation in the new API does not
-        translate paths between operating systems using ftrack disk
+        Proxy the unmanaged location using the legacy API, since the
+        ftrack-python-api does not translate paths between operating systems
+        using ftrack disks.
 
     .. change:: fixed