+++ doc/release/release_notes.rst
@@ -7,6 +7,14 @@
 Release Notes
 *************
 
+.. release:: Upcoming
+    
+    .. change:: new
+
+        Proxy the unmanaged location using the legacy API, since the
+        ftrack-python-api does not translate paths between operating systems
+        using ftrack disks.
+
 .. release:: 0.2.0
     :date: 2017-03-24